+++ packages/create/src/types.ts
@@ -1,4 +1,4 @@
-export type DbType = 'mysql' | 'postgres' | 'sqlite' | 'sqljs' | 'mssql' | 'oracle';
+export type DbType = 'mysql' | 'mariadb' | 'postgres' | 'sqlite' | 'sqljs' | 'mssql' | 'oracle';
 
 export interface UserResponses {
     usingTs: boolean;